Fix for Issue #11

+++ openmp-offload/XSbench_header.h
@@ -70,7 +70,7 @@ typedef struct{
 	int length_concs;
 	int length_mats;
 	int length_unionized_energy_array;
-	int length_index_grid;
+	long length_index_grid;
 	int length_nuclide_grid;
 	int max_num_nucs;
 	double * p_energy_samples;
